+++ src/attribute_info_classes/instruction_classes/push.cpp
@@ -9,7 +9,7 @@ Bipush::Bipush(Attribute::CodeAttribute* code_attr, uint8_t opcode)
 }
 
 int Bipush::execute(Frame* frame){
-    frame->operand_stack.push((int)this->param,Int);
+    frame->operand_stack.push((int8_t)this->param,Int);
 
     return frame->local_pc+=2;
 }
@@ -27,7 +27,7 @@ Sipush::Sipush(Attribute::CodeAttribute* code_attr, uint8_t opcode)
 }
 
 int Sipush::execute(Frame* frame){
-    frame->operand_stack.push((int)this->param,Int);
+    frame->operand_stack.push((int16_t)this->param,Int);
 
     return frame->local_pc+=3;
 }
